Fix CF Workers deploy: break transitive bun:sqlite import chain
Extract portable notification scheduling functions (convertToLocalTime,
computeNotificationCron, refreshNotificationSchedule) into a new
server/jobs/schedule.ts that has no dependency on bun:sqlite. Use a
callback pattern (setScheduleCallback) so the Bun entry point wires up
the actual cron registration while CF Workers gets a no-op.

The transitive chain was: worker.ts → routes/notifiers.ts →
jobs/notifications.ts → jobs/queue.ts → db/bun-db.ts (bun:sqlite).
